model/requirements_new.txt
# ============================================================
# Synapse MD — Python 后端依赖清单（新版）
# 生成日期：2026-04-01
# Python 版本：3.11.15
# 环境：neuroagent (conda)
# 用途：迁移至宝塔 Ubuntu 云端服务器时使用此文件
#
# 与原 requirements.txt 的主要差异：
#   - langchain        0.1.20  → 1.2.14
#   - langchain-community  0.0.38  → 0.4.1
#   - langchain-core   (隐式)  → 1.2.23
#   - langchain-openai 0.1.7   → 1.1.12
#   - langchain-chroma 0.1.4   → 1.1.0
#   - langchain-text-splitters (新增) 1.1.1
#   - langchain_dashscope      (新增) 0.1.8
#   - langgraph        (新增)  1.1.4
#   - chromadb         0.5.23  → 1.5.5  ⚠️ 旧向量库数据不兼容，需删除重建
#   - openai           1.109.1 → 2.30.0 ⚠️ 大版本升级
# ============================================================

# --- Web 框架 ---
fastapi==0.128.0
uvicorn==0.40.0
sse-starlette>=2.1.0      # SSE 协议层保活（ping_interval），替代 asyncio.wait 心跳

# --- LangChain 生态（必须整套协同，不可单独升级某一个）---
langchain==1.2.14
langchain-core==1.2.23
langchain-community==0.4.1
langchain-openai==1.1.12
langchain-chroma==1.1.0
langchain-text-splitters==1.1.1
langchain_dashscope==0.1.8
langchain-classic==1.0.3
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langsmith==0.7.23

# --- 向量库 ---
chromadb==1.5.5           # ⚠️ 从 0.5.23 升级，服务器上需删除旧 chroma_db_unified 目录后重建

# --- AI / LLM SDK ---
dashscope==1.25.9
openai==2.30.0            # ⚠️ 从 1.x 升级到 2.x

# --- 数据处理 ---
pydantic==2.12.5
numpy==1.26.4
pandas==3.0.0
scikit-learn==1.8.0

# --- NLP / 向量模型 ---
transformers==4.38.2
sentence-transformers==2.7.0
torch==2.1.0
rank-bm25==0.2.2
tiktoken==0.7.0

# --- PDF 处理 ---
pdfplumber==0.11.9

# --- 工具类 ---
python-dotenv==1.2.1
PyJWT==2.10.1
requests==2.32.5
httpx==0.28.1
packaging==26.0             # wheel 0.46.3 要求 >=24.0，已升级

# --- 测试 ---
pytest==9.0.2
